Updated Dialogs Positioning/OnlyChildren
Updated the dialogs that are centered to translate3d because there was
text blur when moving. I also changed the center container to be a flexbox
that automatically centers instead. So now there is no css transform trickery
to center it... It is handled by flexbox.

I alos updated the dialogs to have an optional boolean onlyChildren that will
allow you to generate a dialog with only the children given.. So it will not
generate the header and wrappers for you. This is mostly for the picker dialogs.