Release/0.3.x (#126)
* Fixed small bug of adding the transition time to the transform attribute for cross-fade-enter-active

* Fixed typo of exted to extend (#124)

* Updated Version for 0.3.4 Release

This small release just fixed a typo in cards and invalid transition for navigation drawers.

* Updated ListItem shouldComponentUpdate

The ListItem component will no longer use the PureRenderMixin's
shouldComponentUpdate if the component prop is a function. This
will allow the other React component handle the re-rendering.

The main benefit of this is you can now use the activeClassName prop
in react-router's link component and it will update correctly.

* Removed the Unused InlineAutocomplete File

* Fixed Incorrect Documentation on Autocomplete

Fixed the order of the onAutocomplete callback function in the documentation.

* Added clearOnAutocomplete Prop to Autocompletes

The Autocomplete component can now have the value cleared
when auto-completed. This is beneficial if the value of the
autocomplete is not really stored in the text field.

* Updated the Inline Autocompletion for Touch

The Inline Autocomplete now allows for a touch device to autocomplete
when the user touches the suggestion or the text field (when the suggestion is visible).

* Pickers Documentation Update

Added documentation explaining how the Containers are actually
exported instead of the singular `DatePicker || TimePicker`.

* Updated DatePicker Date Format Options

Updated the Date Picker component to allow custom date formats using
Intl. Before this, it would only allow the default configuration of
the date as numbers to be displayed.

* Updated Styling/ClassName SelectField Props

Updated the SelectField to include some additional style props.

* Added an AccessibleFakeButton

There are sometimes you need something to behave like a button, but shouldn't actually
be a button (pretty rare). I added a simple component that adds some accessibility for
those rare times.

Also updated the ListTile to use this new AccessibleFakeButton.

* Added Styling for Tables in Cards

Updated the styling for Tables to match the spects for Tables in Cards.
This can be applied by adding the tableCard prop to the Card component.

* Updated dependencies

* Fixed Custom PropType Validation

Updated the custom PropType validation to include the `...otherUpdated the custom PropType validation to include the `...others`.

* Updated TableRow Hover States

Updated the TableRow hover state to be a className instead of the :hover.
This is because the row should not be highlighted if the mouse is over a menu/menu item
or over an open edit dialog column.

Also updated the styles for Select Fields inside of Tables.

* Small Styling Fixes for Data Tables

- Added the md-table-body className to the TableBody component.
- Fixed the EditDialogColumn's padding so that the text field's value will not be
    extended all the way to the edge of the table. It will keep the table's padding
    correctly.
- Simplified the table-column and table-header padding selectors.
- Fixed the sizes for the sortable icon and the checkboxes.
- Updated the plain data table's base css to be the what I explained it should be
    in the documentation website.

* Added the TablePagination Component

Added a component for helping paginate tables.

* Small margin for for <p> in the CardText component

* Fixed Autocomplete Open Bug

Fixed a bug that the Autocomplete would open for a second if
the list of data changed.

* Updated travis for https://docs.travis-ci.com/user/languages/javascript-with-nodejs#Node.js-v4-(or-io.js-v3)-compiler-requirements

* Fixed Jest Caching for newer version

* I Forgot to add jest-cahce to travis cache..