Portal SSR
Since the portal is really used for dynamic changes, it was unable to be
rendered server side. The _solution_ as of now is to just render the
children as a child in the current tree instead of the subtree. This
will allow it to appear for the weird people that have javascript
disabled but also update itself correctly once working client side.