SelectField Keyboard Accessibility Fixes
Updated the keyboard event listeners for the select field to interact a
bit better. Some of the fixes are:

- No longer need to press the DOWN/UP arrow twice on initial select
field opening when there is no value
- The activeIndex resets correctly when menu is positioned below. This
allows the user to start start from the top again since the active item
is sliced out.
- The SelectField is only refocused when using a keybaord, so touch and
clicks no longer add the ink effect when closing.
- Fixed the bug where you were unable to select the last item when the
menu was positioned below.