Added Dialog Keyboard Closing
Updated the Dialog component to be able to be close when the escape key
is pressed. This can be disabled by passing `closeOnEsc: false` to the
component.

If the Dialog is considered a modal, the escape keypress will still not
do anything and require the user to select one of the actions.