Fixed the Markdown creation
Fixed the  markdown so that it correctly uses the additional languages
and plugins that I need for my site. Also updated the Markdown links to
use the history push state when internal links and behave as normal
links when referencing external sites. In addition, the external site
links are updated to include security rel.