Updated Floating Button Sizing
The floating buttons will now have the same size between disabled and
normal states. This was done by using the paper class names instead of
apply the md-box-shadow mixin to the floating button. This allows for
the floating button sizing to still be applied and toggling the
box-shadow off when disabled.

Closes #295