Added Flexbox Dialog Opt-In Support
Since I was doing an old version of centering elements with a left/top +
transform translate3d for dialogs, it caused the dialogs to be blurry.
It is now possible to fix this by enabling a new boolean
`$md-dialog-use-flex-positioning`. This will remove the old centering
logic and replace it with a flexbox approach.

Some "negative" side-affects is that you will now have to position
dialogs using margin instead of top/right/bottom/left, but it's still
simple.

This should close #337, #599, and #696